ci: force `node-gyp-build` to only find native addons in `prebuilds` in `scripts/pkg.js`
Due to the replacement of `prebuildify` with `scripts/prebuild.js`, the
script no longer moves the native addon to the prebuilds directory, and
instead copies the native addon. This means when using `node-gyp-build`
to find the correct binaries for bundling, it prefers binaries in the
`build/` directory instead of binaries in the `prebuilds/` directory.
This fixes ensures that when using `node-gyp-build` inside
`scripts/pkg.js`, only the prebuild binaries are bundled.