ci: fixing windows build tests by removing `utp-native`, `fd-lock`, `level` and changing back to nodejs 16.14.2
There's a problem on Windows runners on gitlab when running
`tests/bin/typescript-demo-lib.test.ts`.  This may be due to the Windows
runner, the dependencies with native addons, or the NodeJS version.  The
jest program on Windows just crashes with no errors, and no signs.

The only clue is that on Windows, the output log is out of order. This
may be due to Windows `process.stdout` being asynchronous. This will be
fixed in a subsequent work that standardises Windows and POSIX logging
behaviour in `process.stdout` and `process.stderr`.